+++ agents/core_agent.py
@@ -24,10 +24,10 @@ from core.embedding import (
 )
 from core.imgen import generate_image_with_retry_smartgen
 from core.llm import LLMError, call_llm, call_llm_with_tools
+from core.tools.tools import Tools
 from core.voice import speak_text, transcribe_audio
 
 from .tools.default_tool_box import DefaultToolBox
-from .tools.tools import Tools
 from .tools.tools_mcp import Tools as ToolsMCP
 
 # Set up logging
